Add modified query for getting metrics for projects
